backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
